Factor out routes from main app
Main app file was getting too unwieldy
