+++ test/src/service/timestamp_test.py
@@ -1,0 +1,6 @@
+from src.service.timestamp import timestamp
+from conftest import assert_close_to_now
+
+def test_timestamp():
+    ts = timestamp()
+    assert_close_to_now(ts)
